Conect cart front to back

+++ client/src/redux/reducer/index.js
@@ -5,8 +5,8 @@ const initialState = {
   category: [],
   profile: [],
   review: [],
-  cu:[],
-  putUser:[],
+  cu: [],
+  putUser: [],
   //-----------------VER QUE SON ESTOS ESTADOS!
   currentUser: { carts: [{ productCart: [] }] },
   cart: {},
@@ -14,7 +14,12 @@ const initialState = {
   //-------------
   users: [],
   prodCart: localStorage.getItem('prodCart') ? JSON.parse(localStorage.getItem('prodCart')) : [],
-  allreviews:[],
+  cartUser: [],
+  cartUserPRUEBA: [],
+  testStatus: [],
+  //-------------
+  searchedUsers: [],
+  allreviews: [],
 };
 
 const orderProducts = (orderSelected, stateProducts) => {
@@ -35,6 +40,9 @@ const orderProducts = (orderSelected, stateProducts) => {
       return stateProducts;
   }
 };
+
+
+
 function rootReducer(state = initialState, action) {
   switch (action.type) {
     case "GET_PRODUCT_NAME":
@@ -58,7 +66,7 @@ function rootReducer(state = initialState, action) {
       return {
         ...state,
       };
-      case "PUT":
+    case "PUT":
       return {
         ...state,
       };
@@ -70,11 +78,11 @@ function rootReducer(state = initialState, action) {
     case "HIDE_CATEGORY":
       return {
         ...state,
-      };      
+      };
     case "DELETE_REVIEW":
       return {
         ...state,
-      };      
+      };
     case "GET_CAT":
       return {
         ...state,
@@ -143,51 +151,79 @@ function rootReducer(state = initialState, action) {
       return {
         ...state,
       };
-    // case "GET_CART":
-    //   return {
-    //     ...state,
-    //     prodCart: action.payload,
-    //   };
+    case "GET_CART":
+      return {
+        ...state,
+        cartUser: action.payload,
+      };
     // case "DELETE_PRODUCT_CART":
     //   return {
     //     ...state,
     //   };
     case 'ADD_PROD_TO_CART':
       const newProd = action.payload;
-      const itemsInCart = state.prodCart?.find((e) => e.id === newProd.id);
+      const isLogin = action.login;
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////////  
+      console.log('newProd----> ESTE ES EL PRODUCTO A AGREGAR', newProd)
+      ////////////////////////////////////////////////////////////////////////////////////////////////////////////  
+
+      //---------------------------------------SI EL USUARIO TRAE CART, LO TENGO QUE COMPLETAR CON EL ID
+      const itemsUserCart = [];
+
+      if (state.cartUserPRUEBA.length > 0 && state.prodCart.length === 0) {
+        state.cartUserPRUEBA.map((e) => itemsUserCart.push({
+          id: e.ProductId.id,
+          name: e.ProductId.name,
+          image: e.ProductId.image,
+          price: e.ProductId.price,
+          brand: e.ProductId.brand,
+          in_Stock: e.ProductId.in_Stock,
+          CategoryId: e.ProductId.CategoryId,
+          rating: e.ProductId.rating,
+          quantity: e.quantity,
+        }))
+      }
+
+      //console.log('itemsUserCart ---> ESTO ES LO QUE ME TRAIGO DE DB', itemsUserCart)
+
+
+      //---------------------------------------ACA YA ME DOY CUENTA SI TENGO UN CARRITO DEL USUARIO O NO (EL UNICO CASO SERIA QUE EL USUARIO NO TENGO UN CARRITO Y QUE EL LOCALSTORAGE A LA HORA DE LOGEARSE ESTE VACIO, ES DECIR QUE EL USUARIO NO HAYA AGREGADO NADA)
+      const itemsInCart = itemsUserCart.length > 0
+        ? itemsUserCart?.find((e) => e.id === newProd.id)
+        : state.prodCart?.find((e) => e.id === newProd.id)
+
+      //console.log('ESTO DEVUELVE SI ENCONTRO EL MISMO ID(itemsInCart)', itemsInCart)
+
 
       if (newProd.quantity < newProd.in_Stock) {
-        const cartItems = itemsInCart
-          ? state.prodCart.map((it) => it.id === newProd.id
-            ? { ...it, quantity: it.quantity + 1 }
-            : it)
-          : [...state.prodCart, { ...newProd, quantity: 1 }];
-        localStorage.setItem('prodCart', JSON.stringify(cartItems));
+        const cartItems = itemsUserCart.length > 0
+          ? itemsInCart//<---- SI EL USUARIO TIENE ALGO EN SU CARRITO  
+            ? itemsUserCart.map((it) => it.id === newProd.id
+              ? { ...it, quantity: it.quantity + 1 } : it)
+            : [...itemsUserCart, { ...newProd, quantity: 1 }]
+
+          : itemsInCart //<---- SI EL USUARIO NO TIENE NADA EN SU CARRITO      
+            ? state.prodCart.map((it) => it.id === newProd.id
+              ? { ...it, quantity: it.quantity + 1 } : it)
+            : [...state.prodCart, { ...newProd, quantity: 1 }];
+
+        //console.log('ASI QUEDA EL VALOR SI SE ENCONTRO', cartItems)
+
+
+        if (!isLogin) {//SI ESTA LOGEADO QUE NO CARGUE EL LOCALSTORAGE SOLO VA A USAR EL ESTADO
+          localStorage.setItem('prodCart', JSON.stringify(cartItems));//lo borro dentro del handle
+        }
+
         return {
           ...state,
           prodCart: cartItems,
         };
+
       } else {
         return {
           ...state,
         };
       }
-
-
-    // const newProd = action.payload;
-    // const itemsInCart = state.prodCart?.find((e) => e.id === newProd.id);
-    // const cartItems = itemsInCart
-    //   ?.quantity < newProd.in_Stock
-    //   ? state.prodCart.map((it) => it.id === newProd.id
-    //     ? { ...it, quantity: it.quantity + 1 }
-    //     : it)
-    //   : [...state.prodCart, { ...newProd, quantity: 1 }];
-    // localStorage.setItem('prodCart', JSON.stringify(cartItems));
-
-    // return {
-    //   ...state,
-    //   prodCart: cartItems,
-    // };
     case 'REMOVE_PROD_FROM_CART':
       const prodToRemove = state.prodCart?.find((e) => e.id === action.payload.id);
       const cartUpgrade = prodToRemove?.quantity > 1
@@ -195,16 +231,23 @@ function rootReducer(state = initialState, action) {
           ? { ...it, quantity: it.quantity - 1 }
           : it)
         : state.prodCart.filter((upgrade) => upgrade.id !== action.payload.id);
-      cartUpgrade.length > 0
-        ? localStorage.setItem('prodCart', JSON.stringify(cartUpgrade))
-        : localStorage.setItem('prodCart', JSON.stringify([]))
-      console.log('cartUpgrade', cartUpgrade)
 
+      let confirmDelete = action.payload.quantity === 1 ? window.confirm("Do you are sure, to delet all cart?") : true
 
-      return {
-        ...state,
-        prodCart: cartUpgrade,
-      };
+      if (confirmDelete) {
+        cartUpgrade.length > 0
+          ? localStorage.setItem('prodCart', JSON.stringify(cartUpgrade))
+          : localStorage.setItem('prodCart', JSON.stringify([]))
+
+        return {
+          ...state,
+          prodCart: cartUpgrade,
+        };
+      } else {
+        return {
+          ...state,
+        };
+      }
     case 'REMOVE_ALL_PRODUCTS_BYID':
       const prodToRemoveAll = state.prodCart?.filter((e) => e.id !== action.payload.id);
       prodToRemoveAll.length > 0
@@ -225,16 +268,55 @@ function rootReducer(state = initialState, action) {
         ...state,
         cu: action.payload,
       };
+    case 'GET_CART_USER':
+      const cartUserAux = action.payload;
+      if (cartUserAux.length > 0) {
+        const itemsUserCart = [];
+        cartUserAux.map((e) => itemsUserCart.push({
+          id: e.ProductId.id,
+          name: e.ProductId.name,
+          image: e.ProductId.image,
+          price: e.ProductId.price,
+          brand: e.ProductId.brand,
+          in_Stock: e.ProductId.in_Stock,
+          CategoryId: e.ProductId.CategoryId,
+          rating: e.ProductId.rating,
+          quantity: e.quantity,
+        }))
+        console.log('SE LEVANTA EL CARRITO DEL USER Y SE DEBE DE ACTUALIZAR PRODCART')
+        localStorage.removeItem('prodCart')
+        return {
+          ...state,
+          cartUserPRUEBA: cartUserAux,
+          prodCart: itemsUserCart,
+        }
+      } else {
+        return {
+          ...state,
+          cartUserPRUEBA: action.payload,
+        };
+
+      }
+
+    case 'CLEAR_CART_USER':
+      return {
+        ...state,
+        cartUserPRUEBA: [],
+      };
     case 'POST_REVIEW':
       return {
         ...state,
       };
     case 'GET_REVIEW':
-      return{
+      return {
         ...state,
         review: action.payload
-            };
-
+      };
+    // case "GET_CART_USER":
+    //   return {
+    //     ...state,
+    //     pruebaUsers: action.payload,
+    //   };
     default:
       return state;
   }
